+++ src/AudioCore/BeeThree.h
@@ -10,17 +10,17 @@ public:
 
 	BeeThree();
 	~BeeThree();
-	void NoteOn( double frequency, double amplitude );
-	double Tick( unsigned int channel = 0 );
-	AudioFrames& Tick( AudioFrames& frames, unsigned int channel = 0 );
+	void NoteOn( float frequency, float amplitude );
+	float Tick( int channel = 0 );
+	AudioFrames& Tick( AudioFrames& frames, int channel = 0 );
 
 };
 
-inline double BeeThree::Tick( unsigned int ) {
-	double temp;
+inline float BeeThree::Tick( int ) {
+	float temp;
 
-	if ( mod_depth_ > 0.0 )	{
-		temp = 1.0 + ( mod_depth_ * vibrato_.Tick() * 0.1 );
+	if ( mod_depth_ > 0.0f )	{
+		temp = 1.0f + ( mod_depth_ * vibrato_.Tick() * 0.1f );
 		waves_[0]->SetFrequency( base_frequency_ * temp * ratios_[0] );
 		waves_[1]->SetFrequency( base_frequency_ * temp * ratios_[1] );
 		waves_[2]->SetFrequency( base_frequency_ * temp * ratios_[2] );
@@ -28,17 +28,17 @@ inline double BeeThree::Tick( unsigned int ) {
 	}
 
 	waves_[3]->AddPhaseOffset( twozero_.GetLastOut() );
-	temp = control1_ * 2.0 * gains_[3] * adsr_[3]->Tick() * waves_[3]->Tick();
+	temp = control1_ * 2.0f * gains_[3] * adsr_[3]->Tick() * waves_[3]->Tick();
 	twozero_.Tick( temp );
-	temp += control2_ * 2.0 * gains_[2] * adsr_[2]->Tick() * waves_[2]->Tick();
+	temp += control2_ * 2.0f * gains_[2] * adsr_[2]->Tick() * waves_[2]->Tick();
 	temp += gains_[1] * adsr_[1]->Tick() * waves_[1]->Tick();
 	temp += gains_[0] * adsr_[0]->Tick() * waves_[0]->Tick();
-	last_frame_[0] = temp * 0.125;
+	last_frame_[0] = temp * 0.125f;
 	return last_frame_[0];
 }
 
-inline AudioFrames& BeeThree::Tick( AudioFrames& frames, unsigned int channel ) {
-	unsigned int channel_count = last_frame_.GetChannelCount();
+inline AudioFrames& BeeThree::Tick( AudioFrames& frames, int channel ) {
+	int channel_count = last_frame_.GetChannelCount();
 	#if defined(flagDEBUG)
 
 	if ( channel > frames.GetChannelCount() - channel_count ) {
@@ -47,15 +47,15 @@ inline AudioFrames& BeeThree::Tick( AudioFrames& frames, unsigned int channel ) 
 	}
 
 	#endif
-	double* samples = &frames[channel];
-	unsigned int j, step = frames.GetChannelCount() - channel_count;
+	float* samples = &frames[channel];
+	int j, step = frames.GetChannelCount() - channel_count;
 
 	if ( channel_count == 1 ) {
-		for ( unsigned int i = 0; i < frames.GetFrameCount(); i++, samples += step )
+		for ( int i = 0; i < frames.GetFrameCount(); i++, samples += step )
 			* samples++ = Tick();
 	}
 	else {
-		for ( unsigned int i = 0; i < frames.GetFrameCount(); i++, samples += step ) {
+		for ( int i = 0; i < frames.GetFrameCount(); i++, samples += step ) {
 			*samples++ = Tick();
 
 			for ( j = 1; j < channel_count; j++ )
