+++ src/AudioCore/Compressor.h
@@ -35,8 +35,8 @@ class Compressor : public Effect {
 public:
 	Compressor();
 	void Clear() override;
-	double Tick( double input, unsigned int channel ) override;
-	double Tick2( double input1, double input2, unsigned int channel ) override;
+	float Tick( float input, int channel ) override;
+	float Tick2( float input1, float input2, int channel ) override;
 	void LoadState(const ArrayMap<String, Object>& state) override;
 	
 protected:
